+++ src/main/java/org/synyx/urlaubsverwaltung/sicknote/web/SickNoteForm.java
@@ -141,8 +141,7 @@ public class SickNoteForm {
             ", endDate=" + endDate +
             ", dayLength=" + dayLength +
             ", aubStartDate=" + aubStartDate +
-            ", aubEndDate=" + aubEndDate +
-            ", comment='" + comment + '\'' +
+            ", aubEndDate=" + aubEndDate + '\'' +
             '}';
     }
 }