+++ src/test/java/org/synyx/urlaubsverwaltung/account/domain/AccountTest.java
@@ -25,6 +25,6 @@ public class AccountTest {
         final String accountToString = account.toString();
         assertThat(accountToString).isEqualTo("Account{person=Person{id='10'}, validFrom=-999999999-01-01," +
             " validTo=+999999999-12-31, annualVacationDays=10, vacationDays=null, remainingVacationDays=10," +
-            " remainingVacationDaysNotExpiring=10, comment='Comment'}");
+            " remainingVacationDaysNotExpiring=10}");
     }
 }